Add spanNoReceiver helper function for creating spans with no receiver
